asdas

+++ Jenkinsfile
@@ -73,10 +73,15 @@ stages {
             steps
                 { //CLEANUP
                 echo 'Clean system of old images and containers'
-                  sh """docker rm \$(docker stop \$(docker ps -a --filter "label=type=${params.filter}" --format="{{.ID}}"))"""
-                  sh """docker image prune -f --filter \"label=type=${params.filter}\""""
 
-                
+                script
+                {
+                  def nr_cont = sh 'docker ps -a | wc -l'
+                  if ( nr_cont > 1){
+                      sh """docker rm \$(docker stop \$(docker ps -a --filter "label=type=${params.filter}" --format="{{.ID}}"))"""
+                      }
+                 sh """docker image prune -f --filter \"label=type=${params.filter}\""""
+                }            
                 
                 }
              }
@@ -84,8 +89,8 @@ stages {
               {
                steps
                 { //Build image and deploy container
-                echo 'Deploy container'
-                sh """docker build -t ${params.imageName}:${BUILD_NUMBER} """
+                echo 'Deploy container'                   
+                sh """docker build -t ${params.imageName}:${BUILD_NUMBER} ."""
                 sh """docker run -d -p ${params.EXPOSED_PORT}:${params.INSIDE_PORT} --name ${params.containerName} ${params.imageName}:${BUILD_NUMBER}"""
                 }
               }
